agent-service/requirements.txt
# Web Framework
fastapi>=0.108.0
uvicorn[standard]>=0.25.0

# Agent Framework
langgraph
langchain
langchain-anthropic
langsmith

# Async & WebSockets
redis>=5.0.0
websockets>=12.0

# Database
sqlalchemy>=2.0.0
asyncpg
alembic

# Machine Learning
river
numpy
pandas
scikit-learn
scipy

# LLM Providers
anthropic
openai

# Utilities
pydantic>=2.0
pydantic-settings
python-dotenv
httpx
python-jose[cryptography]

# Monitoring
prometheus-client

# Development & Testing
pytest
pytest-asyncio
black
ruff
locust  # For load testing
